requirements.txt
celery==5.2.7
Django==3.2.13
django-allauth==0.54.0
django-filter==23.1
Pillow==9.5.0
redis==4.5.4

